chore: cleanup, update and refactor
