+++ .dockerignore
@@ -1,3 +1,5 @@
 *
-!recycle
+!datacycle
 !pyproject.toml
+!poetry.lock
+!README.md
