+++ .pre-commit-config.yaml
@@ -1,12 +1,12 @@
 repos:
 - repo: https://github.com/pre-commit/pre-commit-hooks
-  rev: v3.4.0
+  rev: v4.2.0
   hooks:
   - id: check-added-large-files
   - id: check-json
   - id: check-yaml
 - repo: https://github.com/sqlalchemyorg/zimports/
-  rev: 0.2.0
+  rev: v0.5.0
   hooks:
   - id: zimports
 - repo: https://github.com/psf/black
@@ -14,10 +14,11 @@ repos:
   hooks:
   - id: black
 - repo: https://gitlab.com/pycqa/flake8
-  rev: 3.8.4
+  rev: 3.9.2
   hooks:
   - id: flake8
-- repo: https://github.com/kynan/nbstripout
-  rev: 0.3.9
+- repo: https://github.com/commitizen-tools/commitizen
+  rev: v2.23.0
   hooks:
-    - id: nbstripout
+  - id: commitizen
+    stages: [commit-msg]
