+++ AUTHORS
@@ -1,4 +1,4 @@
-# This is the list of recycle's significant contributors.
+# This is the list of datacycle's significant contributors.
 #
 # This does not necessarily list everyone who has contributed code,
 # especially since many employees of one corporation may be contributing.
