+++ datacycle/main.py
@@ -4,11 +4,11 @@ from typing import Optional
 
 import dataconf
 from furl import furl
-from recycle import config
-from recycle.providers import gcs
-from recycle.providers import mongo as mongodb
-from recycle.providers import postgres as postgresql
-from recycle.utils import shell
+from datacycle import config
+from datacycle.providers import gcs
+from datacycle.providers import mongo as mongodb
+from datacycle.providers import postgres as postgresql
+from datacycle.utils import shell
 from typer import echo
 from typer import Exit
 from typer import Option
@@ -193,8 +193,19 @@ def doctor():
     shell("gsutil --version")
 
 
+utils = Typer()
+
+@utils.command()
+def gen_gcs_csek():
+    import base64
+    import os
+    echo(base64.b64encode(os.urandom(32)).decode())
+
+
+cli.add_typer(utils, name="utils")
+
 @cli.callback()
 def main():
     """
-    Smood data mover.
+    Datacycle.
     """
