+++ tests/utils.py
@@ -1,6 +1,6 @@
+from datacycle.providers import mongo
+from datacycle.providers import postgres
 from faker import Faker
-from recycle.providers import mongo
-from recycle.providers import postgres
 from sqlalchemy import Boolean
 from sqlalchemy import Column
 from sqlalchemy import DateTime
